refactored mapping creation
Refactoring the mapping creation into a utility class to allow reuse in fully aoutmation commands.
