+++ com.ibm.rm.typemanagement/src/main/java/com/ibm/requirement/typemanagement/oslc/client/resources/DngCmDeliverySession.java
@@ -47,10 +47,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.ibm.requirement.typemanagement.oslc.client.automation.DngTypeSystemManagementConstants;
+import com.ibm.requirement.typemanagement.oslc.client.automation.util.DngHeaderRequestInterceptor;
 import com.ibm.requirement.typemanagement.oslc.client.automation.util.LoggingUtil;
 import com.ibm.requirement.typemanagement.oslc.client.automation.util.RDFUtils;
 import com.ibm.requirement.typemanagement.oslc.client.automation.util.TimeStampUtil;
-import com.ibm.requirement.typemanagement.oslc.client.dngcm.DngHeaderRequestInterceptor;
 import com.ibm.requirement.typemanagement.oslc.client.tracking.IRequestTracker;
 import com.ibm.requirement.typemanagement.oslc.client.tracking.RequestTrackerImpl;
